Disallow default to sqlite if postgres not found
The default setting, which was to spin up an sqlite db if
TEST_DATABASE_URL was not specified caused errors since ARRAY is not
implement in sqlite. Better to be faced with an error that the variable
was not found, than an error that ARRAY is not available to sqlite.