Improve Classified.generate_fake() method
Method was missing a for loop, and hence was only producing a fraction
of the fake classifications.

Previously this method would 'identify' pii 50% of the time. This is
unrealistic, and should be closer to 0.01%. Use numpy.random.count to
apply a probability weighting.

Also ensure that the machine only picks from surveys with priority < 6
like a human classifier.